# openapi.yaml
openapi: 3.0.3
info:
  title: Percolator API
  version: 0.1.0
  description: |
    Percolator is a high-performance perpetual futures trading engine built on Solana.
    This API provides read access to market data, trades, funding rates, and platform statistics.
    
    ## Features
    - Real-time market data and pricing
    - Trade history and volume analytics
    - Funding rate calculations and history
    - Open interest tracking
    - Insurance fund monitoring
    - Platform-wide statistics
    - WebSocket support for live updates
    
    ## Rate Limits
    - Read endpoints: 100 requests/minute
    - Write endpoints: 20 requests/minute
    
    ## Caching
    Most endpoints implement response caching with varying TTLs (10-60 seconds) to optimize performance.
    
  contact:
    name: Percolator Team
    url: https://percolator.trade
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.percolator.trade
    description: Production server
...
paths:
...
  /markets/{slab}/trades:
    get:
      tags:
        - Trades
      summary: Get recent trades for a market
      description: Returns recent trades for a specific market
      operationId: getMarketTrades
      parameters:
        - $ref: '#/components/parameters/SlabAddress'
        - name: limit
          in: query
          description: Maximum number of trades to return (max 200)
          schema:
            type: integer
            default: 100
            maximum: 200
      responses:
        '200':
          description: Successfully retrieved trades
          content:
            application/json:
              schema:
                type: object
                properties:
                  trades:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trade'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /markets/{slab}/volume:
    get:
      tags:
        - Trades
      summary: Get 24h volume for a market
      description: Returns 24-hour trading volume and trade count for a specific market
      operationId: getMarketVolume
      parameters:
        - $ref: '#/components/parameters/SlabAddress'
      responses:
        '200':
          description: Successfully retrieved volume
          content:
            application/json:
              schema:
                type: object
                properties:
                  slab_address:
                    type: string
                  volume_24h:
                    type: string
                    description: Total volume in last 24 hours (as string to preserve precision)
                  trade_count_24h:
                    type: integer
                    description: Number of trades in last 24 hours
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /markets/{slab}/prices:
    get:
      tags:
        - Prices
      summary: Get price history for a market
      description: Returns historical price data for charting (default 24h)
      operationId: getMarketPriceHistory
      parameters:
        - $ref: '#/components/parameters/SlabAddress'
        - name: hours
          in: query
          description: Number of hours of history to retrieve (max 720 = 30 days)
          schema:
            type: integer
            default: 24
            minimum: 1
            maximum: 720
      responses:
        '200':
          description: Successfully retrieved price history
          content:
            application/json:
              schema:
                type: object
                properties:
                  slab_address:
                    type: string
                  prices:
                    type: array
                    items:
                      $ref: '#/components/schemas/PricePoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /trades/recent:
    get:
      tags:
        - Trades
      summary: Get recent trades across all markets
      description: Returns recent trades globally across all markets
      operationId: getGlobalTrades
      parameters:
        - name: limit
          in: query
          description: Maximum number of trades to return (max 200)
          schema:
            type: integer
            default: 100
            maximum: 200
      responses:
        '200':
          description: Successfully retrieved trades
          content:
            application/json:
              schema:
                type: object
                properties:
                  trades:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trade'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /prices/markets:
    get:
      tags:
        - Prices
      summary: Get current prices for all markets
      description: Returns latest price data for all markets
      operationId: getAllMarketPrices
      responses:
        '200':
          description: Successfully retrieved prices
          content:
            application/json:
              schema:
                type: object
                properties:
                  markets:
                    type: array
                    items:
                      type: object
                      properties:
                        slab_address:
                          type: string
                        last_price:
                          type: string
                          nullable: true
                        mark_price:
                          type: string
                          nullable: true
                        index_price:
                          type: string
                          nullable: true
                        updated_at:
                          type: string
                          format: date-time
        '500':
          $ref: '#/components/responses/InternalServerError'
  /prices/{slab}:
    get:
      tags:
        - Prices
      summary: Get price history for a market
      description: Returns the last 100 oracle price updates for a specific market
      operationId: getMarketPrices
      parameters:
        - $ref: '#/components/parameters/SlabAddress'
      responses:
        '200':
          description: Successfully retrieved price history
          content:
            application/json:
              schema:
                type: object
                properties:
                  prices:
                    type: array
                    items:
                      $ref: '#/components/schemas/OraclePrice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /open-interest/{slab}:
    get:
      tags:
        - Open Interest
      summary: Get open interest for a market
      description: Returns current open interest data and historical records (15s cache)
      operationId: getMarketOpenInterest
      parameters:
        - $ref: '#/components/parameters/SlabAddress'
      responses:
        '200':
          description: Successfully retrieved open interest data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenInterestDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /insurance/{slab}:
    get:
      tags:
        - Insurance
      summary: Get insurance fund data
      description: Returns current insurance fund balance and historical data
      operationId: getMarketInsurance
      parameters:
        - $ref: '#/components/parameters/SlabAddress'
      responses:
        '200':
          description: Successfully retrieved insurance data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsuranceDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    SlabAddress:
      name: slab
      in: path
      required: true
      description: Market slab address (Solana public key)
      schema:
        type: string
        pattern: '^[1-9A-HJ-NP-Za-km-z]{32,44}$'
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: string
        slab_address:
          type: string
        side:
          type: string
          enum: [long, short]
        price_e6:
          type: string
        size:
          type: string
        timestamp:
          type: string
          format: date-time
        signature:
          type: string
    PricePoint:
      type: object
      properties:
        price:
          type: number
          description: Price as decimal
        price_e6:
          type: string
          description: Price in e6 format (for precision)
        timestamp:
          type: string
          format: date-time
    OraclePrice:
      type: object
      properties:
        slab_address:
          type: string
        price_e6:
          type: string
        timestamp:
          type: string
          format: date-time
        source:
          type: string
          nullable: true
...
    OpenInterestDetails:
      type: object
      properties:
        slabAddress:
          type: string
        totalOpenInterest:
          type: string
        netLpPos:
          type: string
        lpSumAbs:
          type: string
        lpMaxAbs:
          type: string
        history:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
              totalOi:
                type: string
              netLpPos:
                type: string
    InsuranceDetails:
      type: object
      properties:
        slabAddress:
          type: string
        currentBalance:
          type: string
        feeRevenue:
          type: string
        totalOpenInterest:
          type: string
        history:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
              balance:
                type: string
              feeRevenue:
                type: string
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        hint:
          type: string
          description: Hint for resolving the error
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid slab address
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Market stats not found
            hint: Market may not have been cranked yet or does not exist
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Failed to fetch data
            details: Database connection timeout